General Logger fixes
Use a non-templated version of etl::string

Set the etl format as a static variable for the Logger

This prevents wasting time recreating the format variable every time the
LOG function is called.

Force inlining the LOG function

This forces the compiler to check the `if constexpr` structure, and
disable the logging line entirely if needed.

Never redefine the log level

Move `disabled` logging level to ease level checks