Added test cases for the new functions
Initially I tried copying the core code of the previous functions and only really calling the new functions in the test cases, but apparently some of the things it uses are private and I couldn't get them to work. I tried using the unit_test::Tester struct that I think was made for this but don't think it can work in its current form, and I'm not sure how it should be modified.