Combine the TM[11,10] functions into one and fix naming and documentation
the "handeTM" functions for TC[11,9] and TC[11,16] were combined into one, since they did the same thing but for a different list of activities, which can just be provided by calling "TC" function instead of having two separate ones with the same code. Also changed the naming and documentation to reflect better on the ECSS standard.